backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
sentence-transformers==3.0.0
faiss-cpu==1.13.2
groq==0.11.0
python-dotenv==1.0.1
numpy==1.26.4
pydantic==2.6.4